five hours of pain and debugging later we have started progress on our voronoi-themed deaths

+++ src/navigation/global_planner.h
@@ -35,7 +35,7 @@ public:
     // to int32 for insertion into voronoi_builder vb_.
     // 100x means we keep to 3 decimal places, probably reasonable.
     // potential issue: if (int32) (coords * 100) overflows? very possible, but unlikely.
-    const uint64_t SCALE_FACTOR = 100;
+    const double SCALE_FACTOR = 100;
 
 // GLOBAL VARIABLES
     // goal coords scaled by SCALE_FACTOR
